requirements/_base.txt
Django==1.10.5
unipath==1.1
djangorestframework==3.5.3
ldap3==2.2.0
django-redis==4.7.0
